style: fix all clippy warnings (919 -> 0)
- Auto-fixed 324 machine-applicable warnings (unnecessary_cast, collapsible_if,
  manual_range_contains, identity_op, is_multiple_of, needless_range_loop,
  needless_return, collapsible_match, needless_late_init, etc.)
- Manual fixes for remaining 105 warnings:
  - Added per-item #[allow] for upper_case_acronyms (Bochs naming conventions)
  - Converted loop indexing to iterator patterns across 20+ files
  - Fixed doc comment formatting (empty lines, overindented items, lazy continuation)
  - Replaced manual checked division with .checked_div().unwrap_or()
  - Replaced manual clamp with .clamp()
  - Fixed field_reassign_with_default in pic.rs
  - Added # Safety doc sections for unsafe functions
  - Removed wildcard_in_or_patterns (3|_ -> _)
  - Used copy_from_slice for manual_memcpy
  - Various other lint fixes

+++ rusty_box/src/cpu/avx512_mask.rs
@@ -97,7 +97,7 @@ impl<I: BxCpuIdTrait> BxCpuC<'_, I> {
         let laddr = self.resolve_addr(instr);
         let seg = BxSegregs::from(instr.seg());
         let val = if self.long64_mode() {
-            self.read_virtual_qword_64(seg, laddr as u64)?
+            self.read_virtual_qword_64(seg, laddr)?
         } else {
             self.v_read_dword(seg, laddr)? as u64
         };
@@ -137,7 +137,7 @@ impl<I: BxCpuIdTrait> BxCpuC<'_, I> {
         let laddr = self.resolve_addr(instr);
         let seg = BxSegregs::from(instr.seg());
         if self.long64_mode() {
-            self.write_virtual_qword_64(seg, laddr as u64, val)?;
+            self.write_virtual_qword_64(seg, laddr, val)?;
         } else {
             self.v_write_dword(seg, laddr, val as u32)?;
         }
